Load the policy model from local node storage.
Eagle is having some problems with the high-speed networking and data
storage. When all rollout workers were trying to load the policy model
at the same time, this was causing all of them to stop and wait.
As a temporary fix, I added a script to copy the policy model to the
/tmp/scratch local storage of each of the rollout nodes, and updated
the games to check there first.